+++ packages/cli/src/next/builders/php/resource/index.ts
@@ -12,3 +12,4 @@ export * from './wpPost/identity';
 export * from './wpPost/mutations';
 export * from './wpTaxonomy';
 export * from './wpOption';
+export * from './transient';